backend/requirements.txt
fastapi>=0.103.1,<0.104.0
uvicorn>=0.23.2,<0.24.0
sqlalchemy>=2.0.20,<2.1.0
psycopg2-binary>=2.9.7
pydantic>=2.3.0,<2.4.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.4.0,<8.5.0
httpx>=0.26.0,<0.27.0